rename: ui/common 패키지로 이동
