+++ app/src/main/java/com/cmc15th/pluv/ui/home/migrate/direct/DisplayMigrationPathScreen.kt
@@ -33,8 +33,8 @@ import com.cmc15th.pluv.core.designsystem.component.LoadingDialog
 import com.cmc15th.pluv.core.designsystem.component.TopBarWithProgress
 import com.cmc15th.pluv.core.designsystem.theme.Title1
 import com.cmc15th.pluv.domain.model.PlayListApp
-import com.cmc15th.pluv.ui.contract.GoogleApiContract
-import com.cmc15th.pluv.ui.contract.SpotifyAuthContract
+import com.cmc15th.pluv.ui.common.contract.GoogleApiContract
+import com.cmc15th.pluv.ui.common.contract.SpotifyAuthContract
 import com.cmc15th.pluv.ui.home.getAppNameRes
 import com.cmc15th.pluv.ui.home.migrate.common.component.FetchPlaylistLoadingIcon
 import com.cmc15th.pluv.ui.home.migrate.common.component.PreviousOrMigrateButton
