+++ app/src/main/java/com/cmc15th/pluv/ui/home/migrate/direct/SelectMigrationMusicScreen.kt
@@ -35,8 +35,8 @@ import com.cmc15th.pluv.core.designsystem.theme.Title1
 import com.cmc15th.pluv.core.ui.component.MusicItem
 import com.cmc15th.pluv.core.ui.component.MusicsHeader
 import com.cmc15th.pluv.domain.model.PlayListApp
-import com.cmc15th.pluv.ui.contract.GoogleApiContract
-import com.cmc15th.pluv.ui.contract.SpotifyAuthContract
+import com.cmc15th.pluv.ui.common.contract.GoogleApiContract
+import com.cmc15th.pluv.ui.common.contract.SpotifyAuthContract
 import com.cmc15th.pluv.ui.home.migrate.common.component.PreviousOrMigrateButton
 import com.cmc15th.pluv.ui.home.migrate.common.component.SourceToDestinationText
 import kotlinx.coroutines.delay
